+++ zoomimage-compose-coil-core/src/commonTest/kotlin/com/github/panpf/zoomimage/compose/coil/core/test/CoilZoomStateTest.kt
@@ -7,7 +7,6 @@ import coil3.ImageLoader
 import coil3.PlatformContext
 import com.github.panpf.zoomimage.coil.CoilModelToImageSource
 import com.github.panpf.zoomimage.coil.CoilModelToImageSourceImpl
-import com.github.panpf.zoomimage.compose.rememberZoomImageLogger
 import com.github.panpf.zoomimage.rememberCoilZoomState
 import com.github.panpf.zoomimage.subsampling.ImageSource.Factory
 import com.github.panpf.zoomimage.test.TestLifecycle
@@ -32,17 +31,11 @@ class CoilZoomStateTest {
                     actual = zoomState1.modelToImageSources.joinToString { it::class.qualifiedName!! }
                 )
 
-                val logger = rememberZoomImageLogger("Test")
                 val modelToImageSources = remember {
                     listOf(TestCoilModelToImageSource()).toImmutableList()
                 }
                 val zoomState2 = rememberCoilZoomState(
                     modelToImageSources = modelToImageSources,
-                    logger = logger
-                )
-                assertEquals(
-                    expected = "Test",
-                    actual = zoomState2.logger.tag
                 )
                 assertEquals(
                     expected = listOf(