+++ zoomimage-compose-sketch3-core/src/androidTest/kotlin/com/github/panpf/zoomimage/compose/sketch3/core/test/SketchZoomStateTest.kt
@@ -1,8 +1,6 @@
 package com.github.panpf.zoomimage.compose.sketch3.core.test
 
-import androidx.compose.runtime.remember
 import androidx.compose.ui.test.junit4.createComposeRule
-import com.github.panpf.zoomimage.compose.rememberZoomImageLogger
 import com.github.panpf.zoomimage.rememberSketchZoomState
 import com.github.panpf.zoomimage.test.TestLifecycle
 import org.junit.Rule
@@ -18,17 +16,10 @@ class SketchZoomStateTest {
     fun testRememberSketchZoomState() {
         rule.setContent {
             TestLifecycle {
-                val zoomState1 = rememberSketchZoomState()
+                val zoomState = rememberSketchZoomState()
                 assertEquals(
                     expected = "SketchZoomAsyncImage",
-                    actual = zoomState1.logger.tag
-                )
-
-                val logger = rememberZoomImageLogger("Test")
-                val zoomState2 = rememberSketchZoomState(logger = logger)
-                assertEquals(
-                    expected = "Test",
-                    actual = zoomState2.logger.tag
+                    actual = zoomState.logger.tag
                 )
             }
         }