+++ CHANGELOG.md
@@ -10,6 +10,7 @@ zoom:
   function. [#35](https://github.com/panpf/zoomimage/issues/35)
 * broken: Add the suspend modifier to the \*ToImageSource methods of the CoilModelToImageSource, GlideModelToImageSource, and PicassoDataToImageSource interfaces
 * new: Added support for keyboard scale and drag functions
+* broken: Remove the logger parameter of the 'remember\*ZoomState()' series of functions
 
 ## 1.1.0-alpha05